+++ app/readme-visual-pack/page.tsx
@@ -1,0 +1,5 @@
+import ReadmeVisualPackExport from "../../components/ReadmeVisualPackExport";
+
+export default function ReadmeVisualPackPage() {
+  return <ReadmeVisualPackExport />;
+}